requirements.txt
aiogram==3.3.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
tortoise-orm==0.20.0
asyncpg==0.29.0
redis==5.0.1
httpx==0.26.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-json-logger==2.0.7
python-dotenv==1.0.0
aerich==0.7.2
